requirements.txt
python-dotenv==1.0.1
tenacity==9.0.0
qdrant_client==1.12.2
llama-index-core==0.12.8
llama-index-readers-docling==0.3.0
llama-index-node-parser-docling==0.3.0
llama-index-readers-file==0.4.1
llama-index-embeddings-gemini==0.3.0
llama-index-llms-gemini==0.4.2
llama-index-vector-stores-qdrant==0.4.2
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
colorama==0.4.6